Simplify Research Use Statement section in DAR review Vote tab
